+++ src/cpp/model/gradido/TransactionBody.h
@@ -41,7 +41,7 @@ namespace model {
 				const std::string& memo,
 				const MemoryBin* senderPublicKey,
 				const MemoryBin* recipiantPublicKey,
-				Poco::UInt32 amount,
+				Poco::Int64 amount,
 				TransactionTransferType transferType,
 				const std::string groupAlias = "",
 				Poco::Timestamp pairedTransactionId = Poco::Timestamp()
@@ -50,7 +50,7 @@ namespace model {
 			static Poco::AutoPtr<TransactionBody> create(
 				const std::string& memo,
 				const MemoryBin* recipiantPublicKey,
-				Poco::UInt32 amount,
+				Poco::Int64 amount,
 				Poco::DateTime targetDate
 			);
